actualizando TurnoDTO y agregando Test

+++ proyectoIntegrador1.2/src/main/java/com/example/proyectoIntegrador11/controller/OdontologoController.java
@@ -19,7 +19,7 @@ public class OdontologoController {
     private OdontologoService odontologoService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<Odontologo> buscarPorId(@PathVariable("id") Integer id) {
+    public ResponseEntity<Odontologo> buscarPorId(@PathVariable("id") Long id) {
        Optional<Odontologo> odontologoBuscado = odontologoService.buscarOdontologoPorId(id);
        if (odontologoBuscado.isPresent()) {
            return ResponseEntity.ok(odontologoBuscado.get());
@@ -47,7 +47,7 @@ public class OdontologoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> eliminar(@PathVariable("id") Integer id) throws ResouceNotFounException {
+    public ResponseEntity<String> eliminar(@PathVariable("id") Long id) throws ResouceNotFounException {
         Optional<Odontologo> odontologo = odontologoService.buscarOdontologoPorId(id);
         if (odontologo.isPresent()) {
             odontologoService.eliminarOdontologo(id);
