+++ proyectoIntegrador1.2/src/main/java/com/example/proyectoIntegrador11/controller/PacienteController.java
@@ -20,7 +20,7 @@ public class PacienteController {
     private PacienteService pacienteService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<Paciente> bucarPacienteId(@PathVariable("id") Integer id) {
+    public ResponseEntity<Paciente> bucarPacienteId(@PathVariable("id") Long id) {
         Optional<Paciente> paciente = pacienteService.buscarPacientePorId(id);
         if (paciente.isPresent()) {
             return ResponseEntity.ok(paciente.get());
@@ -29,7 +29,7 @@ public class PacienteController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> eliminarPacienteId(@PathVariable("id") Integer id) throws ResouceNotFounException {
+    public ResponseEntity<String> eliminarPacienteId(@PathVariable("id") Long id) throws ResouceNotFounException {
         Optional<Paciente> paciente = pacienteService.buscarPacientePorId(id);
         if (paciente.isPresent()) {
             pacienteService.eliminarPaciente(id);
